requirements.txt
streamlit==1.31.0
pandas==2.2.3
numpy==1.26.4
sentence-transformers==3.1.1
faiss-cpu==1.9.0
openai==1.51.0
tiktoken==0.8.0
pint==0.24.3
scikit-learn==1.5.2
plotly==5.24.1
python-dotenv==1.0.1